+++ frontend/src/App.tsx
@@ -1,7 +1,7 @@
 import { BrowserRouter, Routes, Route } from "react-router-dom";
 import Navbar from "@/components/ui/Navbar/Navbar";
 import Home from "@/pages/Home/Home";
-import Login from "./components/ui/Login/login";
+import Login from "./components/ui/Login/register";
 import { ThemeToggle } from "./components/theme-toggle";
 
 export default function App() {
